app/requirements.txt
# Minimal requirements for Vercel deployment - ALL DYNAMIC VERSIONS
fastapi
python-dotenv
structlog
openai
anthropic
requests
